+++ HISTORY.rst
@@ -26,6 +26,7 @@ latest (unreleased)
 **Features and Improvements**
 
 * Do not install PLM and Quality modules
+* Zendesk ticket numbers validation
 
 **Bugfixes**